+++ routes/api/contacts.js
@@ -1,5 +1,6 @@
 const express = require("express");
 const ctrl = require("../../controllers/contacts");
+const { postValidation, putValidation } = require("../../middlewares/validationMiddlewares")
 
 const router = express.Router();
 
@@ -7,10 +8,10 @@ router.get("/", ctrl.getAllContacts);
 
 router.get("/:contactId", ctrl.getByIdContacts);
 
-router.post("/", ctrl.postContacts);
+router.post("/", postValidation, ctrl.postContacts);
 
 router.delete("/:contactId", ctrl.deleteContacts);
 
-router.put("/:contactId", ctrl.updateByIdContacts);
+router.put("/:contactId", putValidation, ctrl.updateByIdContacts);
 
 module.exports = router;
